trader_api/requirements.txt
Django==3.2.16
django-cors-headers>=3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
psycopg2==2.9.6